// backend/src/modules/activity/enums/action.enum.ts
export enum ActivityType {
  NOTE_CREATED = 'NOTE_CREATED',
  NOTE_UPDATED = 'NOTE_UPDATED',
  NOTE_DELETED = 'NOTE_DELETED',
  NOTE_PINNED = 'NOTE_PINNED',
  NOTE_UNPINNED = 'NOTE_UNPINNED',
  NOTE_FAVORITED = 'NOTE_FAVORITED',
  NOTE_UNFAVORITED = 'NOTE_UNFAVORITED',
  NOTE_ARCHIVED = 'NOTE_ARCHIVED',
  NOTE_UNARCHIVED = 'NOTE_UNARCHIVED',
}